+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- added additional information in `getInfo()` for `getResponse()` after throw `HttpExceptionInterface` exception
+
 ## 1.0.0 - 2023-05-12
 
 - initial release